requirements.txt
# Core
python-dotenv>=1.0.1
pydantic>=2.8.2
# PDF / text
pypdf>=5.0.0
markdown>=3.6
beautifulsoup4>=4.12.3
# Embeddings & Vector store
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0.post1
# App & eval
streamlit>=1.37.0
tqdm>=4.66.4
tabulate>=0.9.0
# LLM client
openai>=1.40.0
